fix 'premium expires' clock format
